add shape extension

+++ src/examples/java/reference/SpriteSetHitbox.java
@@ -1,4 +1,5 @@
 package reference;
+
 import org.openpatch.scratch.*;
 import org.openpatch.scratch.extensions.recorder.*;
 
@@ -8,12 +9,11 @@ public class SpriteSetHitbox {
     myStage.setDebug(true);
     Sprite mySprite = new Sprite("zeta", "assets/zeta_green_badge.png");
     myStage.add(mySprite);
-    GifRecorder recorder =
-        new GifRecorder("examples/reference/" + this.getClass().getName() + ".gif");
+    GifRecorder recorder = new GifRecorder("examples/reference/" + this.getClass().getName() + ".gif");
     recorder.start();
     myStage.wait(2000);
-    int x[] = {0, 10, 10, 0};
-    int y[] = {0, 0, 10, 10};
+    double x[] = { 0, 10, 10, 0 };
+    double y[] = { 0, 0, 10, 10 };
     mySprite.setHitbox(x, y);
     recorder.stop();
     myStage.wait(2000);
